added authorization for edit and destroy methods in posts controller

+++ app/Http/Controllers/PostsController.php
@@ -84,6 +84,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id != $post->user_id) {
+          return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -119,6 +125,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id != $post->user_id) {
+          return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success', 'Post successfully removed.');
     }
